+++ Studio/src/main.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { Toaster } from "sonner";
 import App from "./App";
 import { ErrorBoundary } from "./components/common/ErrorBoundary";
 import "./index.css";
@@ -8,6 +9,7 @@ ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
     <ErrorBoundary>
       <App />
+      <Toaster position="bottom-right" theme="dark" richColors />
     </ErrorBoundary>
   </React.StrictMode>,
 );